errors: const errors cannot provide id-uniqueness
const errors would have provided the nicest solution, but unfortunately
they cannot provide uniqueness. Therefore it is possible to falsely
identify an error for another error. This cannot happen.

Reverted to (pointers of) struct instances. Fields are private such that
we can return the exact type, without opening up raw construction of
struct values. NewStringError, NewUintError and NewIntError exist to
construct these error types. The idea of the different error types is to
allow for setting the error instance with the identifying part of the
error, e.g. for http this is a status code. Contexts can be used to
provide additional information about a particular occurrence of the
error.